additional features

+++ src/main/java/com/gda/restapi/app/controller/UserController.java
@@ -3,11 +3,14 @@ package com.gda.restapi.app.controller;
 import static com.gda.restapi.app.util.JsonFilterUtility.userWithoutPasswordFilter;
 
 import java.net.URI;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.json.MappingJacksonValue;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +23,8 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import com.gda.restapi.app.exception.InvalidIdFormatException;
+import com.gda.restapi.app.exception.ResourceForbiddenException;
+import com.gda.restapi.app.model.Role;
 import com.gda.restapi.app.model.User;
 import com.gda.restapi.app.service.UserService;
 
@@ -35,8 +40,14 @@ public class UserController {
 	
 	// GET '/users' //
 	@GetMapping("/users")
-	public ResponseEntity<MappingJacksonValue> retrieveAllUsers() {
-		List<User> users = userService.getAll();
+	public ResponseEntity<MappingJacksonValue> retrieveAllUsers(Authentication authentication) {
+		List<User> users = new ArrayList<>();
+		var isAdmin = isAdmin(authentication);
+		
+		if (isAdmin)
+			users = userService.getUsersByRole(Role.USER);
+		else
+			users = userService.getAll();
 		
 		MappingJacksonValue mappingJacksonValue = 
 							userWithoutPasswordFilter(users);
@@ -47,9 +58,14 @@ public class UserController {
 
 	// POST '/users' //
 	@PostMapping("/users")
-	public ResponseEntity<MappingJacksonValue> createUser(@RequestBody @Valid User user) {
+	public ResponseEntity<MappingJacksonValue> createUser(@RequestBody @Valid User user, Authentication authentication) {
 		User savedUser = userService.create(user);
 		
+		var isAdmin = isAdmin(authentication);
+		
+		if (isAdmin && !user.getRole().equals(Role.USER))
+			throw new ResourceForbiddenException("You don't have privilege to access these resources");
+
 		UriComponentsBuilder builder = ServletUriComponentsBuilder.fromCurrentRequestUri();
 		URI location = builder.path("/" + savedUser.getId()).build().toUri();
 		
@@ -62,7 +78,7 @@ public class UserController {
 
 	// GET '/users/{id}' //
 	@GetMapping("/users/{id}")
-	public ResponseEntity<MappingJacksonValue> retrieveUser(@PathVariable String id) {
+	public ResponseEntity<MappingJacksonValue> retrieveUser(@PathVariable String id, Authentication authentication) {
 		int intId = 0;
 		
 		try {
@@ -72,6 +88,11 @@ public class UserController {
 		}
 		
 		User user = userService.getById(intId);
+		var isAdmin = isAdmin(authentication);
+	
+		if (isAdmin && !user.getRole().equals(Role.USER.name()))
+			throw new ResourceForbiddenException("You can't access these resources");
+			
 		
 		MappingJacksonValue mappingJacksonValue = 
 							userWithoutPasswordFilter(user);
@@ -82,7 +103,7 @@ public class UserController {
 
 	// PUT '/users/{id}' //
 	@PutMapping("/users/{id}")
-	public ResponseEntity<MappingJacksonValue> updateUser(@PathVariable String id, @RequestBody @Valid User user) {
+	public ResponseEntity<MappingJacksonValue> updateUser(@PathVariable String id, @RequestBody @Valid User user, Authentication authentication) {
 		int intId = 0;
 		
 		try {
@@ -91,7 +112,14 @@ public class UserController {
 			throw new InvalidIdFormatException("You need to specify the user ID as an integer number");
 		}
 		
+		var registeredUser = userService.getById(intId);
+		var isAdmin = isAdmin(authentication);
+
+		if (isAdmin && !registeredUser.getRole().equals(Role.ADMIN.name()))
+			throw new ResourceForbiddenException("You can't delete these resources");
+		
 		user.setId(intId);
+		user.setRole(registeredUser.getRole());
 		User updatedUser = userService.update(user);
 
 		MappingJacksonValue mappingJacksonValue = 
@@ -103,7 +131,7 @@ public class UserController {
 
 	// DELETE '/users/{id}' //
 	@DeleteMapping("/users/{id}")
-	public ResponseEntity<Void> deleteUser(@PathVariable String id) {
+	public ResponseEntity<Void> deleteUser(@PathVariable String id, Authentication authentication) {
 		int intId = 0;
 		
 		try {
@@ -111,10 +139,24 @@ public class UserController {
 		} catch (NumberFormatException e) {
 			throw new InvalidIdFormatException("You need to specify the user ID as an integer number");
 		}
-		
+
+		var user = userService.getById(intId);
+		var isAdmin = isAdmin(authentication);
+
+		if (isAdmin && !user.getRole().equals(Role.USER.name()))
+			throw new ResourceForbiddenException("You can't delete these resources");
+
 		userService.deleteById(intId);
 
 		return ResponseEntity.noContent().build();
+	}
+
+	
+	private boolean isAdmin(Authentication authentication) {
+		var roles =  authentication.getAuthorities();
+		var isAdmin = roles.contains(
+				new SimpleGrantedAuthority(Role.ADMIN.name()));
+		return isAdmin;
 	}
 
 }
